Wrap publication into a "publisher" process.
The publisher process is a hook later on for the case where you want to
do request/reply patterns to your target. In that case, we can use the publisher
process as the hook point for such messages. Right now, publication is simply
the next step up in the world.