Use the new publisher process over the current setting.
The advantage is that using the publisher process makes the configuration
of the queue static in the supervisor tree. So if you restart the publisher, the
configuration will get re-injected into the system. This ensures that regardless
of where the publishers eventually lives, you can restart it and it should re-obtain
connectivity.