Monitor the connection in the publisher.
If the Publisher dies, then it is restarted. The Channel is then automatically reaped
by RabbitMQ. This is not a problem. If the Connection dies, we install a monitor on
it, so we can track its state. If the connection dies, we stop the publisher with an
error. This ensure we will restart the Publisher by the rules mandated by its tree.