Introduce a test which stops the connection.
Wisdom learned: the RabbitMQ AMQP client is brittle if you start killing random
processes inside it :/. In turn, just killing off stuff will not make it reap itself
correctly, and I managed to link stuff to the wrong connection in the process.

If you gracefully close the connection, everything works out as it should,
and the connection is restarted correctly. However, if you kill the process
which runs the connection, then delivery is happening to the old channel
which in turn means it is never delivered. That is, the state between the Server
and us is wrong, and the server delivers messages with the new delivery tag
on the old channel. This of course fails, because by this point, that old channel
is dead and has no consumers. And we get an unexpected consumer error.

What we are hoping happens here is that errors in RabbitMQ is handled
gracefully by the core of the amqp_client, and that it will close connections
gracefully.