The default AMQP server is called `amqp_server`.
If we use a default name in `turtle` which people will likely use in other projects, we
can exploit this fact to provide simple configurations in projects that use `turtle`.

You can simply request `amqp_server` and you will get the "correct" AMQP server
in the context you are currently operating in, provided the configuration is
correctly setup.