Handle the case where a Sequence No isn't tracked.
If someone is doing a cast on the line, then their SeqNo is not tracked by
the publisher. This is intended, as we have no receiver, so we don't care
about the receiver either.

However, the code assumed there is always a tracking entry, but we have
just shown this to not be the case. Handle the case, and return the empty
list of workers to wake up.