Handle in-flight messages on channels.
If you have a subscriber with prefetch set, then you will have messages
delivered to your mailbox from the channel asynchronously. Suppose that
the subscriber process crashes. These messages in the mailbox is now
gone, but the channel process in the RabbitMQ Erlang Client will *never*
reject these back. This poses a message-leak in turtle.

The fix here is to add a specific handler in terminate/2 for the case. Step
one is to cancel consumption and await the acknowledgement of that
consumption. Next, messages in the mailbox are picked up one by one
and are rejected with requeueing so other processes can handle them.

This fixes the `faulty_service` test case.