Move the service channels onto the subscribers.
This patch alters the internal structure of the turtle subscriber path such
that each subscriber runs on its own channel. It is a necessary change
due to the fact that we want to utilize features which are channel-scoped
in the future, per subscriber. Doing it like this avoids the problem where
multiple consumers on the same channel can accidentally interact with
each other.