Move connection handling onto the janitor process
Creating connections now proceeds by using the janitor process to
manage the connection state. Move monitor references on top of the
janitor as well and push a message to the connection process if the
channel dies. This also ensures connection kills will clean up
connections in the AMQP client tree.