Fix the conn_sup supervision specification
Fixes #34 - The turtle conn process is a worker and not a supervisor.

I have also upgraded the specification itself to the new map-based
supervision specification format. This is also the format used by the
janitor specification a couple of lines above, so I do not think there
is any backwards compatibility concerns.